fix: Fixed production post-index failure bug
Addressed issue causing queries to fail
in production before indexes are built
leading to a 500 error. Now, posts
are sorted using the 'sorted' function
and not via indexes so user-pages
will be viewable before indexes are
completely built. Note missing index
query failure  is an implementation limitation
of Google App Engine and this is a workaround
to this limitation.
See Google App Engine documentation regarding
'Cloud Datastore Indexes' for more info.

Resolves: #10